+++ src/view/direct_mode/direct_channel_widget.py
